refactor: remove redundant, duplicate entity-config constant (#2050)
