+++ src/app/core/entity/model/entity.ts
@@ -36,8 +36,6 @@ export type EntityConstructor<T extends Entity = Entity> = (new (
 ) => T) &
   typeof Entity;
 
-export const ENTITY_CONFIG_PREFIX = "entity:";
-
 /**
  * "Entity" is a base class for all domain model classes.
  * It implements the basic general properties and methods that are required for all Entity types
